chore(release): release 0_0_12 and github action
