+++ package.json
@@ -1,15 +1,18 @@
 {
   "name": "@network-goods/hypercerts-sdk",
-  "version": "0.0.9",
+  "version": "0.0.12",
   "description": "SDK for hypercerts protocol",
   "repository": "git@github.com:Network-Goods/hypercerts-sdk.git",
   "author": "Network Goods",
   "license": "MIT",
+  "source": "src/index.ts",
   "typedocMain": "src/index.ts",
-  "main": "./lib/src/index.js",
-  "types": "./lib/src/index.d.ts",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
   "files": [
-    "lib/**/*"
+    "lib",
+    "README.md",
+    "package.json"
   ],
   "dependencies": {
     "@graphprotocol/client-add-source-name": "latest",
@@ -57,7 +60,7 @@
     "mockttp": "^3.6.1",
     "nyc": "^15.1.0",
     "shx": "^0.3.4",
-    "ts-mocha": "^10.0.0",
+    "ts-jest-resolver": "^2.0.0",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.0"
   },
@@ -65,20 +68,18 @@
     "graphql": "^16.6.0"
   },
   "scripts": {
-    "graph:build": "yarn graphclient build --fileType ts --throwOnInvalidConfig && tsc",
-    "graph:build:client": "graphclient build",
-    "graph:build:esm": "graphclient build --fileType js --throwOnInvalidConfig",
+    "graph:build": "rm -rf ./src/.graphclient && NODE_OPTIONS='--loader ts-node/esm' graphclient build --throwOnInvalidConfig && mv .graphclient ./src/.graphclient",
     "graph:compile": "graphql-codegen",
     "graph:serve": "graphclient serve-dev",
-    "graph:tsc": "tsc",
-    "graph:test": "echo \"Error: No tests\" && exit 1",
     "graph:watch": "graphql-codegen -w",
-    "clean": "rm -rf ./lib",
-    "build": "npm run clean && npm run graph:build:esm && npm run build:esm",
+    "graph:build:esm": "rm -rf ./src/.graphclient && NODE_OPTIONS='--loader ts-node/esm' graphclient build --filetype js --throwOnInvalidConfig && mv .graphclient ./src/.graphclient",
+    "clean": "rm -rf ./lib && rm -rf ./src/types",
+    "build": "yarn clean && yarn types:json && yarn graph:build:esm && yarn build:esm",
     "build:esm": "tsc",
-    "prepack": "npm run build",
+    "prebuild": "yarn clean",
+    "prepack": "yarn build",
     "test": "NODE_OPTIONS=\"--experimental-vm-modules\" jest",
-    "types:json": "yarn run json2ts -i 'resources/schema/' -o 'types' --cwd 'resources/schema' --no-declareExternallyReferenced"
+    "types:json": "yarn run json2ts -i './src/resources/schema/' -o 'src/types' --cwd './src/resources/schema' --no-declareExternallyReferenced"
   },
   "type": "module"
 }
