+++ src/index.ts
@@ -5,19 +5,21 @@ import { storeMetadata, storeData, getMetadata, getData, deleteMetadata } from "
 export { validateMetaData, validateClaimData, storeMetadata, storeData, getMetadata, getData, deleteMetadata };
 
 // Graph
-import { execute } from "../.graphclient/index.js";
+import { execute } from "./.graphclient/index.js";
 import { claimsByOwner, claimById, firstClaims } from "./claims.js";
 import { fractionsByOwner, fractionsByClaim } from "./fractions.js";
 
 export { execute, claimsByOwner, claimById, firstClaims, fractionsByOwner, fractionsByClaim };
 
 // Protocol
-import { HyperCertMinterFactory, HypercertMinterABI } from "@network-goods/hypercerts-protocol";
+import * as protocol from "@network-goods/hypercerts-protocol";
+
+const { HyperCertMinterFactory, HypercertMinterABI } = protocol;
 
 export { HyperCertMinterFactory, HypercertMinterABI };
 
 // Types
-import type { HypercertClaimdata } from "../types/claimdata.js";
-import type { HypercertMetadata } from "../types/metadata.js";
+import type { HypercertClaimdata } from "./types/claimdata.js";
+import type { HypercertMetadata } from "./types/metadata.js";
 
 export type { HypercertMetadata, HypercertClaimdata };
