+++ src/operator/index.ts
@@ -1,10 +1,11 @@
 // @ts-ignore
 import { NFTStorage, CIDString, Blob } from "nft.storage";
-import { HypercertMetadata } from "../../types/metadata.js";
+import { HypercertMetadata } from "../types/metadata.js";
 import axios from "axios";
 
 // NFT.Storage
 const NFT_STORAGE_TOKEN = process.env.NFT_STORAGE_TOKEN ?? "MISSING_TOKEN";
+
 const NFT_STORAGE_IPFS_GATEWAY = "https://nftstorage.link/ipfs/{cid}";
 const defaultNftStorageClient = new NFTStorage({ token: NFT_STORAGE_TOKEN });
 
@@ -21,6 +22,7 @@ export const storeMetadata = async (data: HypercertMetadata, targetClient?: NFTS
 export const getMetadata = async (cid: string): Promise<HypercertMetadata | null> => {
   const nftStorageGatewayLink = NFT_STORAGE_IPFS_GATEWAY.replace("{cid}", cid);
   console.log(`Getting metadata ${cid} at ${nftStorageGatewayLink}`);
+
   return axios
     .get<HypercertMetadata>(nftStorageGatewayLink)
     .then((result) => result.data)
@@ -34,7 +36,7 @@ export const storeData = async (data: any, targetClient?: NFTStorage): Promise<C
   const client = targetClient ?? defaultNftStorageClient;
 
   const blob = new Blob([JSON.stringify(data)], { type: "application/json" });
-  console.log("Storing blob: ", blob);
+  console.log("Storing blob of: ", data);
 
   return await client.storeBlob(blob);
 };
