+++ src/resources/schema/metadata.json
