+++ test/operator.test.ts
@@ -1,25 +1,42 @@
 import { expect, assert } from "chai";
-import { getMetadata, storeMetadata } from "../src/index.js";
-import mockData from "./mockData.json";
+import { getData, getMetadata, storeData, storeMetadata } from "../src/index.js";
+import metadata from "./res/mockMetadata.json";
+import data from "./res/mockData.json";
 
-import { HypercertMetadata } from "../types/metadata.js";
+import { HypercertMetadata } from "../src/types/metadata.js";
 
-const mockMetadata: HypercertMetadata = mockData;
+const mockMetadata: HypercertMetadata = metadata;
+const mockData: any = data;
+
 const mockMetadataCid = "bafkreigdm2flneb4khd7eixodagst5nrndptgezrjux7gohxcngjn67x6u";
+
+const mockDataCid = "bafkreif5otrkydrrjbp532a75hkm5goefxv5rqg35d2wqm6oveht4hqto4";
 
 describe("IPFS Client", () => {
   /**
    * Currently just testing against the production NFT.Storage service.
    */
-  it("Smoke test - add", async () => {
+  it("Smoke test - add metadata", async () => {
     const result = await storeMetadata(mockMetadata);
     expect(result).to.be.a("string");
     expect(result).to.equal(mockMetadataCid);
   });
 
-  it("Smoke test - get", async () => {
+  it("Smoke test - get metadata", async () => {
     const data = await getMetadata(mockMetadataCid);
 
     expect(data).to.deep.equal(mockMetadata);
+  });
+
+  it("Smoke test - add data", async () => {
+    const result = await storeData(mockData);
+    expect(result).to.be.a("string");
+    expect(result).to.equal(mockDataCid);
+  });
+
+  it("Smoke test - get data", async () => {
+    const data = await getData(mockDataCid);
+
+    expect(data).to.deep.equal(mockData);
   });
 });
