+++ test/validator.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import { validateClaimData, validateMetaData } from "../src/validator/index.js";
-import { HypercertMetadata } from "../types/metadata.js";
-import { HypercertClaimdata } from "../types/claimdata.js";
+import { HypercertMetadata } from "../src/types/metadata.js";
+import { HypercertClaimdata } from "../src/types/claimdata.js";
 
 const exampleMetaData: HypercertMetadata = {
   name: "test",
@@ -25,7 +25,6 @@ const exampleClaimData: HypercertClaimdata = {
 };
 
 describe("Validate claim test", () => {
-
   it("checking default metadata", () => {
     expect(validateMetaData(exampleMetaData)).to.be.true;
     expect(validateMetaData({} as HypercertMetadata)).to.be.false;
