+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
   "compilerOptions": {
-    "lib": ["ESNext", "DOM"],
+    "lib": ["ES6", "DOM"],
     "target": "ES6",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
     "baseUrl": ".",
-    "outDir": "./lib/",
+    "outDir": "./lib",
     "allowJs": true,
     "allowSyntheticDefaultImports": true,
     "declaration": true,
@@ -22,5 +22,5 @@
     "skipLibCheck": true
   },
   "exclude": ["node_modules"],
-  "include": ["src/**/*", ".graphclient/**/*"]
+  "include": ["src/**/*", ".graphclient/**/*", "src/.graphclient/sources/**/*"]
 }
